+++ src/tests/EntityGraphQL.Tests/SchemaTests/GraphQLSchemaGenerateTests.cs
@@ -315,11 +315,11 @@ namespace EntityGraphQL.Tests
         {
             var schemaProvider = SchemaBuilder.FromObject<AbstractClassTestSchema>(false);
 
-            schemaProvider.AddType<AbstractClassTestSchema.Dog>("Dogs are animals").AddAllBaseTypes().AddAllFields();
-            schemaProvider.AddType<AbstractClassTestSchema.Cat>("Cats are animals").AddBaseType<Animal>().AddAllFields();
+            schemaProvider.AddType<AbstractClassTestSchema.Dog>("Dogs are animals").ImplementAllBaseTypes().AddAllFields();
+            schemaProvider.AddType<AbstractClassTestSchema.Cat>("Cats are animals").Implements<Animal>().AddAllFields();
             schemaProvider.AddType<AbstractClassTestSchema.Fish>("Fish are animals");
 
-            schemaProvider.UpdateType<AbstractClassTestSchema.Fish>(x => x.AddBaseType("Animal").AddAllFields());
+            schemaProvider.UpdateType<AbstractClassTestSchema.Fish>(x => x.Implements("Animal").AddAllFields());
 
             var schema = schemaProvider.ToGraphQLSchemaString();
             // this exists as it is not null
@@ -336,9 +336,9 @@ namespace EntityGraphQL.Tests
             var schemaProvider = SchemaBuilder.FromObject<AbstractClassTestSchema>(false);
 
             schemaProvider.AddType<AbstractClassTestSchema.ISwim>("").AddAllFields();
-            schemaProvider.AddType<AbstractClassTestSchema.Dog>("Dogs are animals").AddAllBaseTypes().AddAllFields();
-            schemaProvider.AddType<AbstractClassTestSchema.Cat>("Cats are animals").AddBaseType<Animal>().AddAllFields();
-            schemaProvider.AddType<AbstractClassTestSchema.Fish>("Fish are animals").AddAllBaseTypes().AddAllFields();
+            schemaProvider.AddType<AbstractClassTestSchema.Dog>("Dogs are animals").ImplementAllBaseTypes().AddAllFields();
+            schemaProvider.AddType<AbstractClassTestSchema.Cat>("Cats are animals").Implements<Animal>().AddAllFields();
+            schemaProvider.AddType<AbstractClassTestSchema.Fish>("Fish are animals").ImplementAllBaseTypes().AddAllFields();
 
             var schema = schemaProvider.ToGraphQLSchemaString();
             // this exists as it is not null
@@ -458,10 +458,6 @@ namespace EntityGraphQL.Tests
     {
         public List<Animal> Animals { get; set; }
 
-        public abstract class Animal
-        {
-            public string Name { get; set; }
-        }
 
         public interface ISwim
         {