+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Breaking changes:
 - Cleaned up `SchemaType` constructors - using `GqlTypeEnum` instead of many boolean flags
 - Removed obsolete `SchemaProvider.AddInheritedType<TBaseType>`
 - Removed the instance parameter from `AddMutationsFrom` and friends. Mutation "controllers" are now always created per request like an asp.net controller. Use DI for any constructor parameters
+- Renamed `ISchemaType.AddBaseType` to `ISchemaType.Implements` to align with GraphQL language
+  - `ISchemaType.Implements` will throw an exception if you try to implement a non-interface
+- Renamed `ISchemaType.AddAllBaseTypes` to `ISchemaType.ImplementAllBaseTypes` to align with GraphQL language
 
 Changes: